feat(linter): add rule eslint/new-cap
